# src/frontend/shared/openapi/wnav-openapi.yaml
openapi: "3.1.0"
info:
  title: WNAV API
  version: 1.0.0
  description: |
    作業ナビゲーションシステム REST API。terminal-api（ポート 8080）と master-api（ポート 8081）の統合仕様。
    エラーは RFC 9457 Problem Details 形式で返却する。
  contact:
    name: WNAV Engineering
    email: engineering@wnav.example.com
servers:
  - url: http://localhost:8080/api/v1
    description: terminal-api ローカル開発
  - url: http://localhost:8081/api/v1
    description: master-api ローカル開発
  - url: https://wnav-terminal.factory.local/api/v1
    description: terminal-api 本番
  - url: https://wnav-master.factory.local/api/v1
    description: master-api 本番
security:
  - bearerAuth: []
...
paths:
...
  /work-executions:
    post:
      tags: [work-executions]
      operationId: postWorkExecution
      summary: 作業実行開始（API-work-execs-001）
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkExecutionRequest' }
      responses:
        '201': { description: 作成成功, content: { application/json: { schema: { $ref: '#/components/schemas/WorkExecutionEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /work-executions/{id}/suspend:
    post:
      tags: [work-executions]
      operationId: postWorkExecutionSuspend
      summary: 作業中断（API-work-execs-003）
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SuspendWorkExecutionRequest' }
      responses:
        '200': { description: 中断成功, content: { application/json: { schema: { $ref: '#/components/schemas/SuspensionEnvelope' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /work-executions/{id}/resume:
    post:
      tags: [work-executions]
      operationId: postWorkExecutionResume
      summary: 作業再開（API-work-execs-004）
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResumeWorkExecutionRequest' }
      responses:
        '200': { description: 再開成功, content: { application/json: { schema: { $ref: '#/components/schemas/WorkExecutionEnvelope' } } } }
  /work-executions/{id}/complete:
    post:
      tags: [work-executions]
      operationId: postWorkExecutionComplete
      summary: 作業完了（API-work-execs-005）
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteWorkExecutionRequest' }
      responses:
        '200': { description: 完了成功, content: { application/json: { schema: { $ref: '#/components/schemas/WorkExecutionEnvelope' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 で署名された JWT を Bearer トークンとして送信する
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid, description: UUID v7 形式 }
...
  responses:
    Unauthorized:
      description: 認証失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Forbidden:
      description: 認可失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    Conflict:
      description: 競合
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    ValidationError:
      description: バリデーション失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
...
    ResponseMeta:
      type: object
      required: [request_id, server_time, api_version]
      properties:
        request_id: { type: string, format: uuid }
        server_time: { type: string, format: date-time }
        api_version: { type: string, enum: [v1] }
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        error_id: { type: string }
        violations:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
              value: {}
...
    CreateWorkExecutionRequest:
      type: object
      required: [work_order_id, operator_id, device_id, start_timestamp_client]
      properties:
        work_order_id: { type: string, format: uuid }
        operator_id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        start_timestamp_client: { type: string, format: date-time }
    WorkExecution:
      type: object
      properties:
        id: { type: string, format: uuid }
        work_order_id: { type: string, format: uuid }
        operator_id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        status: { type: string, enum: [in_progress, suspended, completed, cancelled] }
        current_step_id: { type: string, format: uuid, nullable: true }
        completed_step_count: { type: integer }
        total_step_count: { type: integer }
        started_at: { type: string, format: date-time }
        last_event_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        hash_chain_block_id: { type: string, format: uuid, nullable: true }
        hash_chain_value: { type: string, nullable: true }
        sop_version_snapshot:
          type: object
          properties:
            sop_id: { type: string, format: uuid }
            version: { type: string }
            snapshot_hash: { type: string }
    WorkExecutionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/WorkExecution' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    SuspendWorkExecutionRequest:
      type: object
      required: [reason_code, timestamp_client]
      properties:
        reason_code: { type: string, enum: [equipment_breakdown, material_shortage, quality_issue, emergency, other] }
        reason_detail: { type: string }
        timestamp_client: { type: string, format: date-time }
    ResumeWorkExecutionRequest:
      type: object
      required: [resumed_by, timestamp_client]
      properties:
        resumed_by: { type: string, format: uuid }
        timestamp_client: { type: string, format: date-time }
    CompleteWorkExecutionRequest:
      type: object
      required: [completed_by, timestamp_client]
      properties:
        completed_by: { type: string, format: uuid }
        timestamp_client: { type: string, format: date-time }
        final_remarks: { type: string }
    SuspensionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          properties:
            id: { type: string, format: uuid }
            status: { type: string }
            suspension_id: { type: string, format: uuid }
            suspended_at: { type: string, format: date-time }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...